+++ xscope_fileio/lib_build_info.cmake
@@ -1,0 +1,6 @@
+set(LIB_NAME xscope_fileio)
+set(LIB_VERSION 1.1.2)
+set(LIB_INCLUDES api .)
+set(LIB_DEPENDENT_MODULES "")
+
+XMOS_REGISTER_MODULE()