+++ xscope_fileio/xscope_io_common.h
@@ -1,11 +1,12 @@
-// Copyright (c) 2020, XMOS Ltd, All rights reserved
+// Copyright 2020-2024 XMOS LIMITED.
+// This Software is subject to the terms of the XMOS Public Licence: Version 1.
 #ifndef XSCOPE_IO_COMMON_H_
 #define XSCOPE_IO_COMMON_H_
 
 #define MAX_XSCOPE_SIZE_BYTES   256
 #define END_MARKER_STRING       "finally_the_end!!" //17 * 8 = 136 bits of unlikely data 
 #define END_MARKER_LEN          (sizeof(END_MARKER_STRING) - 1)
-#define MAX_FILENAME_LEN        64
+#define MAX_FILENAME_LEN        128
 #define MAX_FILES_OPEN          32
 
 typedef enum{
@@ -23,6 +24,7 @@ enum{
     XSCOPE_ID_SEEK          = 4,
     XSCOPE_ID_TELL          = 5,
     XSCOPE_ID_HOST_QUIT     = 6, 
+    XSCOPE_ID_HOST_CLOSE    = 7,   
 };
 
 #endif